Merge pull request from GHSA-4852-vrh7-28rf with some adjustments.
The fork of this package is not used or suggested to be used in the same way
as the upstream package.  The built-in technique within Apollo Server does
not allow users to configure playground in a dynamic per-request manner and
we do not publish anything besides `@apollographql/graphql-playground-html`
and `@apollographql/graphql-playground-react`.  Both are intended for direct
use by Apollo Server.

Co-authored-by: Rikki Schulte <rikki.schulte@gmail.com>