Don't rely on safe-listed properties unnecessarily.
There's not a compelling reason why we can't just put the DIV we want to
preserve outside of the filtering as static prefixes and suffixes.  This
avoids unnecessary complexity and surface area to a security-conscious piece
of code.  For example, without this change any number of unescaped DIVs with
ID properties can be used inside of the escaped content unnecessarily,
merely because we were trying to preserve a DIV that we control.